Remove MtgJsonCardData rapidjson::Value ref member
MtgJsonCardData kept a reference to a rapidjson Value to look up its
information as needed for each method call.  This was fine when the
MtgJsonAllSetsData object's lifetime was forever, but the need to
reload and update the card database may be necessary.

Modified the class to use the rapidjson value to create local copies of
member data at construction time.