Add connection monitoring and client keep-alives
When clients disappeared from the network without an explicit disconnect
the server would not clean up, leaving rooms and usernames dangling for
a long time (maybe indefinitely).

On the server side, ClientConnection now keeps a timer of when it last
heard from a client.  If it's been more than 30 seconds the connection
is aborted.

The client also keeps a timer.  If it hasn't sent data in more than 25
seconds it sends a newly introduced message type (KeepAliveInd) to the
server.