Rework client file save dialog to force non-native
The file save dialog used the static API from Qt which resulted in a
native dialog on Windows.  The native dialog halts the application event
loop, which halts the QTimer that triggers keepalives being sent to the
server, which causes the server to disconnect the client.  Reworked to
use a non-native dialog and keep the event loop running.