Heavily refactor python sim client
Moved Tx/Rx lower-layer protocol stuff to a separate module.  Made the
client into a class of its own.  It now creates rooms if none are
present or joins an existing one.  The client marks itself ready once in
a room and chooses the first card in each pack it receives.

As the first crack at a server stress test, the outer script portion
starts 100 clients with a little delay in each one.