Add 'auto' zone to protocol, server, and client
Added ZONE_AUTO to support autoselected cards in booster draft and
looking ahead to sealed draft.

In this initial version, the client immediately moves autoselected cards
to the former destination zone and sends an inventory update to the server.
In this way the outward-facing behavior should be no different than before.