Improve client/server protocol mismatch handling
Client now gives additional information to server at login time so that
the server may reject the client based on protocol.  When this happens
the server may include a description and URL in the LoginRsp

Client's protocol mismatch dialogs are clearer and in the case of a
server login rejection based on protocol, include the server's
instructions for upgrade.

Updated protocol to include:
  - Client protocol version and software version in LoginReq
  - Server reject code for client protocol version error
  - New optional message in LoginRsp with server instructions for
    obtaining new client software

Protocol version bumped to 1.2.