Add hide-when-empty behavior to chosen tabs
CommanderPane tabs can now be hidden automatically once they become
empty.  If a tab with this setting is currently showing while it
becomes empty, it does not hide but waits until the user navigates away.

Hide-when-empty is enabled in the client for the Auto and Draft tabs.
When the draft is active the behavior is overriden so that the Draft tab
will always be shown.