Add web services and client check for updates
Introduced node.js web service code that listens on port 53332 and
presents APIs for checking for updates and serves static files.

Added client startup and menu-driven checking for updates via access to
the web service API.

Client protocol mismatching is now handled by pointing to a redirecting
link presented by the node.js script.

Added a developer-only command-line option to point the client to a
locally-running instance of the web service on localhost.