Add check/update of MTG JSON card data by channel
Storing current version of AllSets.json and using web services to check
for updates on one of two update channels: stable and mtgjson.  The
stable channel is thicket's web service method, the mtgjson channel uses
mtgjson.com's API.

Currently defaulted to the stable channel.  Channel selection is
available in underlying settings but settable via GUI yet.  At this time
there is no need for end users to use the mtgjson channel because the
server will always be on the stable channel and there's no value in
the client being more up to date than the server.